Adjust \Nova\Net\Session and move \Nova\Helpers\TableBuilder to \Nova\Database\Builder
